refactor: fold type tests into result.php and exclude private ctor from coverage
Review feedback:
- tests/Types/results.php duplicated the role of the existing
  tests/Types/result.php (same namespace, same purpose); fold the
  three assertType functions into the existing file per the
  edit-over-create convention
- the never-executed private constructor was the one uncovered diff
  line failing the codecov patch (92.85% < 100%) and project checks;
  mark it @codeCoverageIgnore
